refactor(interpretation): split overlay.rs into overlay/ submodule
Addresses Nico's #886 review ask (break overlay.rs down; extract graph/link
utilities). The 1034-line overlay.rs becomes overlay/{mod,classes,gate,write}.rs:
- classes.rs — hard-wired InterpretationRun/Overlay SDNA + run mint + registration
- gate.rs    — the §4 human-divergence gate (ownership decision + graph reads)
- write.rs   — overlay persistence (mint-or-bump + inferred/<p> links + presence)
- mod.rs     — apply_with_overlay orchestration + shared vocab (+ #[cfg(test)] seed_overlay)
graph/write.rs absorbs the extracted link/relation helpers. Behaviour-preserving.
